Rename test files to class-test-*.php
